requirements.txt
websockets==12.0
aiohttp==3.10.11
python-dotenv==1.0.0
pydantic==2.10.6
